refactor: extract sqlite crate
