+++ crates/monotax/src/init.rs
@@ -3,7 +3,7 @@
 use sqlx::SqlitePool;
 
 use crate::config::dir::create_default_config;
-use crate::infra::sqlx::init::initialize_db;
+use monotax_sqlite::init::initialize_db;
 
 pub async fn init(db_pool: &SqlitePool, force: bool) -> anyhow::Result<()> {
     create_default_config(force)?;
