+++ crates/monotax_core/Cargo.toml
@@ -9,7 +9,7 @@ license-file = "LICENSE"
 
 [dependencies]
 anyhow = "1.0.81"
-chrono = "0.4.37"
+chrono = "0.4"
 clap = { version = "4.5.4", features = ["derive"] }
 csv = "1.3.0"
 directories = "5.0.1"
