+++ crates/monotax_core/src/infra/mod.rs
@@ -4,4 +4,3 @@
 //! store data and results.
 
 pub mod io;
-pub mod sqlx;
