+++ crates/monotax_core/src/integration/taxer.rs
@@ -6,14 +6,26 @@
 use std::io::Write;
 
 use csv::Writer;
+use serde::{Deserialize, Serialize};
 
-use crate::config::TaxerImportConfig;
 use crate::domain::Income;
 
 pub struct TaxerIncome<'a> {
     income: &'a Income,
     tax_number: &'a str,
     comment: &'a str,
+}
+
+/// Configuration for the Taxer import.
+///
+/// - The `id` is person's national tax identifier.
+/// - The `account_name` is the name of the account in the Taxer.
+/// - The `default_comment` is a comment that will be used if the income has no comment.
+#[derive(Debug, Serialize, Deserialize)]
+pub struct TaxerImportConfig {
+    id: String,
+    account_name: String,
+    default_comment: String,
 }
 
 /// Export incomes to a CSV file that can be imported into Taxer.
@@ -58,5 +70,32 @@ impl<'a> TaxerIncome<'a> {
         let amount = format!("{:.2}", self.income.amount());
         writer.write_record([self.tax_number, date.as_str(), &amount, self.comment])?;
         Ok(())
+    }
+}
+
+impl TaxerImportConfig {
+    /// Provide the national tax identifier.
+    pub fn id(&self) -> &str {
+        &self.id
+    }
+
+    /// Provide the name of the account in the Taxer.
+    pub fn account_name(&self) -> &str {
+        &self.account_name
+    }
+
+    /// Provide the default comment that will be used if the income has no comment.
+    pub fn default_comment(&self) -> &str {
+        &self.default_comment
+    }
+}
+
+impl Default for TaxerImportConfig {
+    fn default() -> Self {
+        Self {
+            id: "1234567890".to_string(),
+            account_name: Default::default(),
+            default_comment: Default::default(),
+        }
     }
 }
