+++ crates/monotax_sqlite/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "monotax_sqlite"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+monotax_core = { path = "../monotax_core/" }
+
+anyhow = "1.0.95"
+async-trait = "0.1.81"
+chrono = "0.4"
+serde = { version = "1", features = ["derive"] }
+sqlx = { version = "0.7", features = [
+  "macros",
+  "sqlite",
+  "runtime-tokio",
+  "chrono",
+] }
+
+[dev_dependencies]
+tokio = { version = "1", features = ["rt", "rt-multi-thread", "macros"] }
+uuid = { version = "1.9.1", features = ["v4"] }
