+++ crates/monotax_sqlite/src/connection.rs
@@ -1,7 +1,7 @@
 //! Connect to the database specified in configuration.
 use sqlx::SqlitePool;
 
-use crate::config::DatabaseConfiguration;
+use crate::configuration::DatabaseConfiguration;
 
 pub async fn connection_pool(config: &DatabaseConfiguration) -> anyhow::Result<SqlitePool> {
     SqlitePool::connect(&config.connection_url())
