+++ crates/monotax_sqlite/src/record.rs
@@ -1,7 +1,7 @@
 use chrono::{Datelike, NaiveDateTime};
 use sqlx::prelude::FromRow;
 
-use crate::domain::{Income, Quarter};
+use monotax_core::domain::{Income, Quarter};
 #[derive(Debug, Clone, FromRow)]
 pub struct IncomeRecord {
     pub date: NaiveDateTime,
